Moved ensureAdcCredentials to own file
